hooks: pre-command: specify path where the binary must be downloaded to
Instead of download to a directory, download the binary, which is by
default of the form `ytt-os-arch[(.exe)?]`, to `DOWNLOAD_DESTINATION/ytt`
itself.
This will ensure that users can directly use `ytt` by simply running
`ytt`, instead of something unnecessarily complicated like
`ytt-linux-amd64`, etc, which is also not very portable.

Also, for the `curl` command, remove the `-z` option, since releases for
`ytt` are well defined and the resource to be downloaded must be
retrieved from GitHub releases that were made in fixed points in time.

Fixes: c36ca0f ("hooks: set up `pre-command`")